issue #87 refactor: sign up profile component separate
