+++ src/components/SignUp/Profile/ProfileForm/ProfileAddress.tsx
@@ -1,0 +1,53 @@
+import RequiredIcon from "@/assets/svg/RequiredIcon.svg?react";
+
+import { Flex, Box, Text } from "@/components/common";
+
+import { getDefaultTextStyle } from "@/styles/getDefaultTextStyle";
+import { Theme } from "@/styles/Theme";
+
+import { addressInputStyle } from "@/components/SignUp/Profile/SignUpProfile.style";
+import { commonButtonStyle, getFormTextStyle } from "@/components/SignUp/SignUp.shared.style";
+
+const ProfileAddress = ({
+	userUrl,
+	changeUserUrl,
+	userUrlRef,
+}: {
+	userUrl: string;
+	changeUserUrl: React.Dispatch<React.SetStateAction<string>>;
+	userUrlRef: React.RefObject<HTMLInputElement>;
+}) => {
+	return (
+		<Flex styles={{ direction: "column", gap: "8px" }}>
+			<Flex styles={{ gap: "4px", align: "center" }}>
+				<Text css={getFormTextStyle(false)}>프로필 주소</Text>
+				<RequiredIcon />
+			</Flex>
+			<Flex styles={{ align: "center", gap: "6px" }}>
+				<Text css={getDefaultTextStyle(Theme.color.text, 500)}>https://www.waggle.com/users/@</Text>
+
+				<input
+					css={addressInputStyle}
+					placeholder="나만의 프로필 주소를 만들어보세요"
+					value={userUrl}
+					onChange={(e) => changeUserUrl(e.target.value)}
+					ref={userUrlRef}
+				/>
+			</Flex>
+
+			<Flex styles={{ align: "center", gap: "16px" }}>
+				<Box tag="button" css={commonButtonStyle}>
+					<Text css={getDefaultTextStyle(Theme.color.disabled_text, 500)}>
+						프로필 주소 중복 확인
+					</Text>
+				</Box>
+
+				<Text css={getDefaultTextStyle(Theme.color.brand_primary, 500)}>
+					사용할 수 있는 주소입니다
+				</Text>
+			</Flex>
+		</Flex>
+	);
+};
+
+export default ProfileAddress;
