+++ src/components/SignUp/Profile/ProfileForm/index.tsx
@@ -1,0 +1,8 @@
+import Birthday from "@/components/SignUp/Profile/ProfileForm/Birthday";
+import Name from "@/components/SignUp/Profile/ProfileForm/Name";
+import Nickname from "@/components/SignUp/Profile/ProfileForm/Nickname";
+import NicknameDescription from "@/components/SignUp/Profile/ProfileForm/NicknameDescription";
+import Profile from "@/components/SignUp/Profile/ProfileForm/Profile";
+import ProfileAddress from "@/components/SignUp/Profile/ProfileForm/ProfileAddress";
+
+export { Profile, Nickname, NicknameDescription, ProfileAddress, Name, Birthday };
