+++ src/components/SignUp/Profile/SignUpProfile.tsx
@@ -2,48 +2,33 @@ import { useState, useReducer, useRef } from "react";
 import { useNavigate } from "react-router-dom";
 import { toast } from "react-toastify";
 
-import DefaultProfileImg from "@/assets/png/profile.png";
-import PasswordCheckIcon from "@/assets/svg/ic-password-check.svg?react";
-import SelectArrowIcon from "@/assets/svg/ic-select-arrow.svg?react";
-import RequiredIcon from "@/assets/svg/RequiredIcon.svg?react";
-
 import { Flex, Box, Text } from "@/components/common";
+import {
+	Profile,
+	Nickname,
+	NicknameDescription,
+	ProfileAddress,
+	Name,
+	Birthday,
+} from "@/components/SignUp/Profile/ProfileForm";
 
-import { yearData, monthData, dayData } from "@/constants/auth";
-import { ALLOW_FILE_EXTENSION, FILE_SIZE_MAX_LIMIT } from "@/constants/file";
 import { SIGN_UP_TAB_KEY, TAB_KEY } from "@/constants/tab";
 
-import { useCheckNicknameMutation } from "@/hooks/api/useCheckNicknameMutation";
 import { useMemberInfoFirstMutation } from "@/hooks/api/useMemberInfoFirstMutation";
 
-import { getDefaultTextStyle } from "@/styles/getDefaultTextStyle";
 import { Theme } from "@/styles/Theme";
 
 import { dateFormatToUTC } from "@/utils/dateFormatToUTC";
-import { fileExtensionValid } from "@/utils/file";
 import { findEmailReducer, findEmailInitialState } from "@/utils/findEmailUtils";
 
-import {
-	imgStyle,
-	getNicknameTextStyle,
-	addressInputStyle,
-	getSelectBoxStyle,
-} from "@/components/SignUp/Profile/SignUpProfile.style";
-import {
-	inputNoneDisplayStyle,
-	commonButtonStyle,
-	getFormTextStyle,
-	getInputStyle,
-	getNextButtonStyle,
-} from "@/components/SignUp/SignUp.shared.style";
+import { getNextButtonStyle } from "@/components/SignUp/SignUp.shared.style";
 
 const SignUpProfile = () => {
-	const checkNicknameMutation = useCheckNicknameMutation();
 	const { mutateMemberInfo } = useMemberInfoFirstMutation();
 
-	const navigate = useNavigate();
+	const [state] = useReducer(findEmailReducer, findEmailInitialState);
 
-	const [state, dispatch] = useReducer(findEmailReducer, findEmailInitialState);
+	const navigate = useNavigate();
 
 	const nicknameRef = useRef<HTMLInputElement>(null);
 	const userUrlRef = useRef<HTMLInputElement>(null);
@@ -52,10 +37,8 @@ const SignUpProfile = () => {
 	const [nickname, setNickname] = useState("");
 	const [userUrl, setUserUrl] = useState("");
 	const [name, setName] = useState("");
-
 	const [fileURL, setFileURL] = useState<string>("");
 
-	const [isNicknameCheck, setIsNicknameCheck] = useState(false);
 	const [nicknameCheckComplete, setNicknameCheckComplete] = useState(false);
 
 	const [userUrlCheckComplete] = useState(false);
@@ -97,50 +80,6 @@ const SignUpProfile = () => {
 		}
 
 		return true;
-	};
-
-	const handleNicknameCheck = () => {
-		setIsNicknameCheck(true);
-
-		checkNicknameMutation.mutate(nickname, {
-			onSuccess: () => {
-				setNicknameCheckComplete(true);
-			},
-		});
-	};
-
-	const handleOptionText = (e: React.MouseEvent<HTMLLIElement>) => {
-		const innerText = e.currentTarget.innerText;
-
-		dispatch({ type: `CHANGE_${e.currentTarget.ariaLabel}_TEXT`, payload: innerText });
-		dispatch({ type: `CHANGE_${e.currentTarget.ariaLabel}_OPTION` });
-		dispatch({ type: `SELECT_${e.currentTarget.ariaLabel}` });
-	};
-
-	const handleChangeImg = (e: React.ChangeEvent<HTMLInputElement>) => {
-		const target = e.currentTarget;
-		const files = (target.files as FileList)[0];
-
-		if (files === undefined) {
-			return;
-		}
-
-		if (!fileExtensionValid(files)) {
-			target.value = "";
-
-			toast.error(`업로드 가능한 확장자가 아닙니다. [가능한 확장자 : ${ALLOW_FILE_EXTENSION}]`);
-
-			return;
-		}
-
-		if (files.size > FILE_SIZE_MAX_LIMIT) {
-			target.value = "";
-			toast.error("업로드 가능한 최대 용량은 1MB입니다. ");
-			return;
-		}
-
-		const newFileURL = URL.createObjectURL(files);
-		setFileURL(newFileURL);
 	};
 
 	const handleSubmit = (e: React.FormEvent) => {
@@ -187,171 +126,28 @@ const SignUpProfile = () => {
 			>
 				<Flex styles={{ direction: "column", gap: "36px" }}>
 					{/* 프로필 영역 */}
-					<Flex styles={{ align: "center", gap: "60px" }}>
-						<img src={fileURL ? fileURL : DefaultProfileImg} alt="profileImg" css={imgStyle} />
-						<Flex styles={{ direction: "column", gap: "14px" }}>
-							<Text css={getFormTextStyle(false)}>프로필 이미지</Text>
-							<input
-								type="file"
-								id="profileImg"
-								onChange={handleChangeImg}
-								css={inputNoneDisplayStyle}
-							/>
-							<label htmlFor="profileImg" css={commonButtonStyle}>
-								<Text css={getDefaultTextStyle(Theme.color.disabled_text, 500)}>
-									컴퓨터에서 파일 선택
-								</Text>
-							</label>
-							<Text size="small" css={getDefaultTextStyle(Theme.color.disabled_text, 500)}>
-								확장자: png, jpg, jpeg / 용량: 1MB 이하
-							</Text>
-						</Flex>
-					</Flex>
+					<Profile fileURL={fileURL} changeFile={setFileURL} />
 
 					{/* 닉네임 영역 */}
-					<Flex styles={{ direction: "column", gap: "8px" }}>
-						<Flex styles={{ gap: "4px", align: "center" }}>
-							<Text css={getFormTextStyle(true)}>닉네임</Text>
-							<RequiredIcon />
-						</Flex>
-						<input
-							css={getInputStyle("444px")}
-							placeholder="닉네임을 입력해주세요! 언제든지 변경 가능해요"
-							value={nickname}
-							onChange={(e) => setNickname(e.target.value)}
-							ref={nicknameRef}
-						/>
-						<Flex styles={{ gap: "16px", align: "center" }}>
-							<Box tag="button" css={commonButtonStyle} onClick={handleNicknameCheck}>
-								<Text css={getDefaultTextStyle(Theme.color.disabled_text, 500)}>
-									닉네임 중복 확인
-								</Text>
-							</Box>
-
-							<Text css={getNicknameTextStyle(isNicknameCheck && nicknameCheckComplete)}>
-								{isNicknameCheck && !nicknameCheckComplete && "사용할 수 없는 닉네임입니다"}
-
-								{isNicknameCheck && nicknameCheckComplete && "사용할 수 있는 닉네임입니다"}
-							</Text>
-						</Flex>
-					</Flex>
+					<Nickname
+						nickname={nickname}
+						changeNickname={setNickname}
+						nicknameRef={nicknameRef}
+						nicknameCheckComplete={nicknameCheckComplete}
+						changeNicknameCheckComplete={setNicknameCheckComplete}
+					/>
 
 					{/* 닉네임 용도 설명 영역 */}
-					<Flex styles={{ direction: "column", gap: "12px" }}>
-						<Text css={getDefaultTextStyle(Theme.color.text, 600)}>닉네임은 이런 곳에 쓰여요!</Text>
-						<Flex styles={{ direction: "column", gap: "8px" }}>
-							<Flex styles={{ align: "center", gap: "6px" }}>
-								<PasswordCheckIcon />
-								<Text size="small" css={getDefaultTextStyle(Theme.color.text, 500)}>
-									닉네임으로 게시물을 작성하거나 댓글을 남길 수 있어요
-								</Text>
-							</Flex>
-							<Flex styles={{ align: "center", gap: "6px" }}>
-								<PasswordCheckIcon />
-								<Text size="small" css={getDefaultTextStyle(Theme.color.text, 500)}>
-									상대방이 나를 언급할 때 내 닉네임을 사용해 태그할 수 있어요
-								</Text>
-							</Flex>
-						</Flex>
-					</Flex>
+					<NicknameDescription />
 
 					{/* 프로필 주소 영역 */}
-					<Flex styles={{ direction: "column", gap: "8px" }}>
-						<Flex styles={{ gap: "4px", align: "center" }}>
-							<Text css={getFormTextStyle(false)}>프로필 주소</Text>
-							<RequiredIcon />
-						</Flex>
-						<Flex styles={{ align: "center", gap: "6px" }}>
-							<Text css={getDefaultTextStyle(Theme.color.text, 500)}>
-								https://www.waggle.com/users/@
-							</Text>
-							<input
-								css={addressInputStyle}
-								placeholder="나만의 프로필 주소를 만들어보세요"
-								value={userUrl}
-								onChange={(e) => setUserUrl(e.target.value)}
-								ref={userUrlRef}
-							/>
-						</Flex>
-
-						<Flex styles={{ align: "center", gap: "16px" }}>
-							<Box tag="button" css={commonButtonStyle}>
-								<Text css={getDefaultTextStyle(Theme.color.disabled_text, 500)}>
-									프로필 주소 중복 확인
-								</Text>
-							</Box>
-							<Text css={getDefaultTextStyle(Theme.color.brand_primary, 500)}>
-								사용할 수 있는 주소입니다
-							</Text>
-						</Flex>
-					</Flex>
+					<ProfileAddress userUrl={userUrl} changeUserUrl={setUserUrl} userUrlRef={userUrlRef} />
 
 					{/* 이름 영역 */}
-					<Flex styles={{ direction: "column", gap: "8px" }}>
-						<Flex styles={{ gap: "4px", align: "center" }}>
-							<Text css={getFormTextStyle(true)}>이름(실명)</Text>
-							<RequiredIcon />
-						</Flex>
-						<input
-							css={getInputStyle("444px")}
-							placeholder="이름을 입력해주세요"
-							value={name}
-							onChange={(e) => setName(e.target.value)}
-							ref={nameRef}
-						/>
-					</Flex>
+					<Name name={name} changeName={setName} nameRef={nameRef} />
 
 					{/* 생년월일 영역 */}
-					<Flex styles={{ direction: "column", gap: "8px" }}>
-						<Flex styles={{ gap: "4px", align: "center" }}>
-							<Text css={getFormTextStyle(true)}>생년월일</Text>
-							<RequiredIcon />
-						</Flex>
-						<Flex styles={{ gap: "13px" }}>
-							{/* 생년 */}
-							<Box css={getSelectBoxStyle(state.year, state.yearSelect)}>
-								<Text onClick={() => dispatch({ type: "CHANGE_YEAR_OPTION" })}>
-									{state.yearText}
-								</Text>
-								<ul>
-									{yearData.map((data) => (
-										<li key={data.selectText} onClick={handleOptionText} aria-label="YEAR">
-											{data.selectText}
-										</li>
-									))}
-								</ul>
-								<SelectArrowIcon />
-							</Box>
-
-							{/* 월 선택 */}
-							<Box css={getSelectBoxStyle(state.month, state.monthSelect)}>
-								<Text onClick={() => dispatch({ type: "CHANGE_MONTH_OPTION" })}>
-									{state.monthText}
-								</Text>
-								<ul>
-									{monthData.map((data) => (
-										<li key={data.selectText} onClick={handleOptionText} aria-label="MONTH">
-											{data.selectText}
-										</li>
-									))}
-								</ul>
-								<SelectArrowIcon />
-							</Box>
-
-							{/* 일 선택 */}
-							<Box css={getSelectBoxStyle(state.day, state.daySelect)}>
-								<Text onClick={() => dispatch({ type: "CHANGE_DAY_OPTION" })}>{state.dayText}</Text>
-								<ul>
-									{dayData.map((data) => (
-										<li key={data.selectText} onClick={handleOptionText} aria-label="DAY">
-											{data.selectText}
-										</li>
-									))}
-								</ul>
-								<SelectArrowIcon />
-							</Box>
-						</Flex>
-					</Flex>
+					<Birthday />
 				</Flex>
 			</Box>
